--- job02_preprocessing.py
@@ -1,77 +1,0 @@
-# loading libraries
-import matplotlib.pyplot as plt
-import numpy as np
-import skimage
-from skimage import measure
-from skimage.transform import radon
-from skimage.transform import probabilistic_hough_line
-from skimage import measure
-from scipy import interpolate
-from scipy import stats
-
-# illustration of 13 regions
-an = np.linspace(0, 2*np.pi, 100)
-plt.plot(2.5*np.cos(an), 2.5*np.sin(an))
-plt.axis('equal')
-plt.axis([-4, 4, -4, 4])
-plt.plot([-2.5, 2.5], [1.5, 1.5])
-plt.plot([-2.5, 2.5], [0.5, 0.5 ])
-plt.plot([-2.5, 2.5], [-0.5, -0.5 ])
-plt.plot([-2.5, 2.5], [-1.5,-1.5 ])
-
-plt.plot([0.5, 0.5], [-2.5, 2.5])
-plt.plot([1.5, 1.5], [-2.5, 2.5])
-plt.plot([-0.5, -0.5], [-2.5, 2.5])
-plt.plot([-1.5, -1.5], [-2.5, 2.5])
-plt.title(" Devide wafer map to 13 regions")
-plt.xticks([])
-plt.yticks([])
-plt.show()
-
-
-def cal_den(x):
-    return 100 * (np.sum(x == 2) / np.size(x))
-
-
-def find_regions(x):
-    rows = np.size(x, axis=0)
-    cols = np.size(x, axis=1)
-    ind1 = np.arange(0, rows, rows // 5)
-    ind2 = np.arange(0, cols, cols // 5)
-
-    reg1 = x[ind1[0]:ind1[1], :]
-    reg3 = x[ind1[4]:, :]
-    reg4 = x[:, ind2[0]:ind2[1]]
-    reg2 = x[:, ind2[4]:]
-
-    reg5 = x[ind1[1]:ind1[2], ind2[1]:ind2[2]]
-    reg6 = x[ind1[1]:ind1[2], ind2[2]:ind2[3]]
-    reg7 = x[ind1[1]:ind1[2], ind2[3]:ind2[4]]
-    reg8 = x[ind1[2]:ind1[3], ind2[1]:ind2[2]]
-    reg9 = x[ind1[2]:ind1[3], ind2[2]:ind2[3]]
-    reg10 = x[ind1[2]:ind1[3], ind2[3]:ind2[4]]
-    reg11 = x[ind1[3]:ind1[4], ind2[1]:ind2[2]]
-    reg12 = x[ind1[3]:ind1[4], ind2[2]:ind2[3]]
-    reg13 = x[ind1[3]:ind1[4], ind2[3]:ind2[4]]
-
-    fea_reg_den = []
-    fea_reg_den = [cal_den(reg1), cal_den(reg2), cal_den(reg3), cal_den(reg4), cal_den(reg5), cal_den(reg6),
-                   cal_den(reg7), cal_den(reg8), cal_den(reg9), cal_den(reg10), cal_den(reg11), cal_den(reg12),
-                   cal_den(reg13)]
-    return fea_reg_den
-
-
-df_withpattern['fea_reg'] = df_withpattern.waferMap.apply(find_regions)
-x = [9, 340, 3, 16, 0, 25, 84, 37]
-labels2 = ['Center', 'Donut', 'Edge-Loc', 'Edge-Ring', 'Loc', 'Random', 'Scratch', 'Near-full']
-
-fig, ax = plt.subplots(nrows=2, ncols=4, figsize=(20, 10))
-ax = ax.ravel(order='C')
-for i in range(8):
-    ax[i].bar(np.linspace(1, 13, 13), df_withpattern.fea_reg[x[i]])
-    ax[i].set_title(df_withpattern.failureType[x[i]][0][0], fontsize=15)
-    ax[i].set_xticks([])
-    ax[i].set_yticks([])
-
-plt.tight_layout()
-plt.show()
